+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 # Next
 
+* Added ability to customize `JSONSerialization.ReadingOptions` for `JSONDecodableSerializer`. Unlike previous releases, no options are specified by default(Previously SwiftyJSON used `allowFragments` option).
+
+To have old behavior that allows fragmented json, use:
+
+```swift
+let request = tron.swiftyJSON(readingOptions: .allowFragments).request("path")
+```
+
+* Added ability to customize `JSONDecoder` for `CodableSerializer` for both model parsing and error parsing.
+* `CodableParser` and `JSONDecodableParser` are now classes, that are subclassible.
+
 ## [4.0.0-beta.1](https://github.com/MLSDev/TRON/releases/tag/4.0.0-beta.1)
 
 **This is major release, containing breaking API changes, please read [TRON 4.0 Migration Guide](https://github.com/MLSDev/TRON/blob/master/Docs/4.0%20Migration%20Guide.md)**